# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAER301 API Documentation",
        "description": "Auto-generated API documentation for SAE301 project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/registrations/{registration}/runners": {
            "get": {
                "tags": [
                    "Race Participants"
                ],
                "summary": "Get registration runners",
                "description": "Returns list of all runners for a specific race registration with PPS information",
                "operationId": "242bc7cf7ff8483946b9c1d3d68738fe",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "Registration ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of runners with PPS status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "runners": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "participant_id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "has_licence": {
                                                        "type": "boolean"
                                                    },
                                                    "licence_number": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "pps_number": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "pps_expiry": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "nullable": true
                                                    },
                                                    "pps_status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "pending",
                                                            "verified",
                                                            "rejected"
                                                        ]
                                                    },
                                                    "has_valid_credentials": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Race Participants"
                ],
                "summary": "Add runner to registration",
                "description": "Add a new runner to a race registration with optional PPS information",
                "operationId": "39f38599989937d5afdced2714506179",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "Registration ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "pps_number": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "nullable": true
                                    },
                                    "pps_expiry": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Runner added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "participant": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User already participating"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/participants/{participant}": {
            "put": {
                "tags": [
                    "Race Participants"
                ],
                "summary": "Update runner PPS information",
                "description": "Update PPS number, expiry date, or status for a race participant",
                "operationId": "fdcc3a95fbd04afad8f27a33d22ad9d0",
                "parameters": [
                    {
                        "name": "participant",
                        "in": "path",
                        "description": "Race Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pps_number": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "nullable": true
                                    },
                                    "pps_expiry": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "pps_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "verified",
                                            "rejected"
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "PPS updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Race Participants"
                ],
                "summary": "Remove runner from registration",
                "description": "Remove a participant from a race registration",
                "operationId": "ef9c64faecce62c102403cfcf73f635b",
                "parameters": [
                    {
                        "name": "participant",
                        "in": "path",
                        "description": "Race Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Runner removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/participants/{participant}/verify-pps": {
            "post": {
                "tags": [
                    "Race Participants"
                ],
                "summary": "Verify runner PPS",
                "description": "Verify or reject a runner's PPS documentation (admin/race manager only)",
                "operationId": "568152104bf2d53c852c9d799bdd5161",
                "parameters": [
                    {
                        "name": "participant",
                        "in": "path",
                        "description": "Race Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Defaults to 'verified' if not provided",
                                        "type": "string",
                                        "enum": [
                                            "verified",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "PPS verification status updated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or race manager only"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}